+++ src/main/java/UniFest/global/infra/security/jwt/exception/AccessTokenExpiredException.java
@@ -1,0 +1,10 @@
+package UniFest.global.infra.security.jwt.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class AccessTokenExpiredException extends FilterException {
+    public AccessTokenExpiredException()
+    {
+        super(HttpStatus.UNAUTHORIZED, "만료된 액세스 토큰입니다", 2000);
+    }
+}
