+++ src/main/java/UniFest/global/infra/security/jwt/exception/RefreshTokenExpiredException.java
@@ -1,4 +1,4 @@
-package UniFest.global.infra.security.jwt.exceptoin;
+package UniFest.global.infra.security.jwt.exception;
 
 import UniFest.global.common.exception.UnifestCustomException;
 import org.springframework.http.HttpStatus;
